+++ ios/brave-ios/Sources/Shared/Reachability.swift
@@ -1,3 +1,8 @@
+// Copyright 2025 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
 import Combine
 import Foundation
 import Network
